Allow using the native LoRA loader node
